Reordering functions related to article modes
