+++ src/clients/artikkelit/actions/articleReset.js
@@ -3,7 +3,7 @@ import {resetAndHideCcLicense} from '/artikkelit/interfaces/articleBasic.js';
 import {resetPublicationSearchResultSelect} from '/artikkelit/interfaces/publicationSearch.js';
 import {resetReviewSearchResultSelect} from '/artikkelit/interfaces/reviewSearch.js';
 import {idbClear, getTableNames} from '/artikkelit/utils/indexedDB.js';
-import {showSnackbar} from '/common/ui-utils.js';
+import {disableElement, enableElement, showSnackbar} from '/common/ui-utils.js';
 
 
 export function idbClearAllTables() {
@@ -42,3 +42,29 @@ function resetSelectFields() {
     selectField.dispatchEvent(new Event('change'));
   }
 }
+
+
+//---------------------------------------------------------------------------//
+// Function for resetting record action buttons and related form and record notes
+export function resetCheckAndSave() {
+  const formNotes = document.getElementById('articleFormNotes');
+  const recordNotes = document.getElementById('articleRecordNotes');
+  const checkArticleRecordButton = document.getElementById('actionCheckArticleRecord');
+  const saveArticleRecordButton = document.getElementById('actionSaveArticleRecord');
+  const forwardIcon = document.getElementById('actionForward');
+
+  formNotes.classList.remove('record-error');
+  formNotes.classList.remove('record-valid');
+  formNotes.classList.remove('record-success');
+  formNotes.innerHTML = '';
+
+  recordNotes.classList.remove('record-error');
+  recordNotes.classList.remove('record-valid');
+  recordNotes.classList.remove('record-success');
+  recordNotes.innerHTML = 'Tarkista tietue ennen tallentamista.';
+
+  disableElement(saveArticleRecordButton);
+  enableElement(checkArticleRecordButton);
+  forwardIcon.classList.remove('proceed');
+}
+
